BlueCarbon/requirements.txt
numpy
pandas
Pillow
matplotlib
seaborn
scikit-learn
tensorflow==2.12.0
fastapi
uvicorn
python-multipart
protobuf
pip install rasterio numpy matplotlib geopandas